requirements.txt
pandas==2.2.2
numpy==1.26.4
sentence-transformers==2.7.0
faiss-cpu==1.8.0
scikit-learn==1.4.2
tqdm==4.66.4